Add new port x11/xorgproto
Upstream has bundled all *proto packages into this meta package.
The plan is in time to have everything that depends on an individual
*proto package depend on this instead.

Some notes:
I have not enabled legacy proto files.  We need to evaluate if we need
those, or if, for the few ports that use them, they can still depend on
the old separate ports.  This can be enabled later with --enable-legacy.
For now, that option is broken upstream, but that can be fixed if we
need it.