SC2035: Use ./*glob* or -- *glob* so names with dashes won't become options.
Since files and arguments are strings passed the same way, programs
can't properly determine which is which, and rely on dashes to determine
what's what.

A file named -f (touch -- -f) will not be deleted by the problematic
code. It will instead be interpreted as a command line option, and rm
will even report success.

Using ./* will instead cause the glob to be expanded into ./-f, which no
program will treat as an option.

It is not possible to use `-f *` because -f only forces the next
argument to be a directory, a later directory named -delete would mess
things up.

PR:             227109
Submitted by:   mat
Sponsored by:   Absolight