x11/xorgproto: Fix build with LEGACY
Fix build with LEGACY option.  Add missing man page for Xprint(7) as a
patch.  For some reason this is missing from the distfile even if it's
in upstream git repo.
Fix plist in LEGACY case.
Enable LEGACY by default, it's needed for several other packages.

Note: I'm not sure if making LEGACY an option is a good idea, since a
lot of other ports need it to be on.  We might want to revisit this
later.