Add devel/evdev-proto and switch consumers over
Add devel/evdev-proto with evdev headers from Linux, instead of using
those in multimeda/v4l_compat.  This makes it easier for the Graphics
Team to keep the headers updated as needed, and they're decoupled from
the rest of the headers in multimedia/v4l_compat.

This is a prerequisite for further updates to the input stack in ports.