Fix xorg module check to look for trailing garbage
Fix the xorg module check to look for trailing garbage after the
optional ':'.  Currently, we only support :both, implying that a module
should be both a build and run time dependency.

There is no need to support :run and :build, since most modules are
shared libraries and needed both at build- and runtime.  The few modules
that aren't shared libraries are always needed at build time, and has no
need for a :run to change that, but can be made a build- and runtime
dependency with :both.  There are no run time only modules that could
use a :build.

As a side note. fontutil (x11-fonts/font-util) is declared as a libray
module, but does not install a shared library.  It should probably be
made a build time dependency instead.