Update based on comments from mat
Change GL_SITE and GL_ACCOUNT around a bit in the case we fetch things
from gitlab.
Add an IGNORE in the meson case, until this can be worked on some more.
Change from MASTER_SITES and MASTER_SITE_SUBDIR to just MASTER_SITES.
In the big block where we check categories, change from a .if/.endif
style for each category, to a .if/.elif/.endif style  It is only
possible to specify one category anyway.

Move checks for USE_XORG earlier, per style.