libpciaccess: Work around driver detection issues
Work around driver detection issues in libpciaccess by removing the
function that detects if a kernel driver is attached to a PCI device.
This causes issues on FreeBSD since the vgapci device is always attached
to the graphics card.
A proper fix is to look if a driver is attached to any child device of
vgapci, but work around the issue in the meantime.

FreeBSD PR:	239065